Conflicts:
	Excel FinancialBI/Excel FinancialBI/AddinModule.Designer.cs
	Excel FinancialBI/Excel FinancialBI/AddinModule.cs
	Excel FinancialBI/Excel FinancialBI/MVC/View/Controlling UI/ResultView.cs
	Excel FinancialBI/Excel FinancialBI/MVC/View/Facts Edition/AFactEditionView.cs

+++ Excel FinancialBI/Excel FinancialBI/MVC/Model/CRUD/EntityDistribution.cs
@@ -13,7 +13,7 @@ namespace FBI.MVC.Model.CRUD
     public UInt32 Id { get; private set; }
     public UInt32 EntityId { get; set; }
     public UInt32 AccountId { get; set; }
-    public uint Percentage { get; set; }
+    public double Percentage { get; set; }
     public UInt32 Image { get; set; }
 
     public EntityDistribution() { }
@@ -28,7 +28,7 @@ namespace FBI.MVC.Model.CRUD
 
       l_entity.EntityId = p_packet.ReadUint32();
       l_entity.AccountId = p_packet.ReadUint32();
-      l_entity.Percentage = p_packet.ReadUint8();
+      l_entity.Percentage = p_packet.ReadDouble();
 
       return (l_entity);
     }
@@ -39,7 +39,7 @@ namespace FBI.MVC.Model.CRUD
         p_packet.WriteUint32(Id);
       p_packet.WriteUint32(EntityId);
       p_packet.WriteUint32(AccountId);
-      p_packet.WriteUint8((byte)Percentage);
+      p_packet.WriteDouble(Percentage);
     }
 
     public void CopyFrom(EntityDistribution p_model)
